requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
pandas>=2.0.0
openpyxl>=3.1.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
baidu-aip>=1.6.6